Move durable nonce to system program
