+++ cli/src/nonce.rs
@@ -10,14 +10,14 @@ use solana_sdk::{
     account::Account,
     account_utils::State,
     hash::Hash,
-    nonce_instruction::{
-        create_nonce_account, nonce_advance, nonce_authorize, nonce_withdraw, NonceError,
-    },
-    nonce_program,
     nonce_state::NonceState,
     pubkey::Pubkey,
     signature::{Keypair, KeypairUtil},
-    system_instruction::SystemError,
+    system_instruction::{
+        create_nonce_account, nonce_advance, nonce_authorize, nonce_withdraw, NonceError,
+        SystemError,
+    },
+    system_program,
     transaction::Transaction,
 };
 
@@ -300,7 +300,7 @@ pub fn check_nonce_account(
     nonce_authority: &Pubkey,
     nonce_hash: &Hash,
 ) -> Result<(), Box<CliError>> {
-    if nonce_account.owner != nonce_program::ID {
+    if nonce_account.owner != system_program::ID {
         return Err(CliError::InvalidNonce(CliNonceError::InvalidAccountOwner).into());
     }
     let nonce_state: NonceState = nonce_account
@@ -407,7 +407,7 @@ pub fn process_create_nonce_account(
 
 pub fn process_get_nonce(rpc_client: &RpcClient, nonce_account_pubkey: &Pubkey) -> ProcessResult {
     let nonce_account = rpc_client.get_account(nonce_account_pubkey)?;
-    if nonce_account.owner != nonce_program::id() {
+    if nonce_account.owner != system_program::id() {
         return Err(CliError::RpcRequestError(format!(
             "{:?} is not a nonce account",
             nonce_account_pubkey
@@ -469,7 +469,7 @@ pub fn process_show_nonce_account(
     use_lamports_unit: bool,
 ) -> ProcessResult {
     let nonce_account = rpc_client.get_account(nonce_account_pubkey)?;
-    if nonce_account.owner != nonce_program::id() {
+    if nonce_account.owner != system_program::id() {
         return Err(CliError::RpcRequestError(format!(
             "{:?} is not a nonce account",
             nonce_account_pubkey
@@ -804,14 +804,14 @@ mod tests {
         let valid = Account::new_data(
             1,
             &NonceState::Initialized(NonceMeta::new(&nonce_pubkey), blockhash),
-            &nonce_program::ID,
+            &system_program::ID,
         );
         assert!(check_nonce_account(&valid.unwrap(), &nonce_pubkey, &blockhash).is_ok());
 
         let invalid_owner = Account::new_data(
             1,
             &NonceState::Initialized(NonceMeta::new(&nonce_pubkey), blockhash),
-            &system_program::ID,
+            &Pubkey::new(&[1u8; 32]),
         );
         assert_eq!(
             check_nonce_account(&invalid_owner.unwrap(), &nonce_pubkey, &blockhash),
@@ -820,7 +820,7 @@ mod tests {
             ))),
         );
 
-        let invalid_data = Account::new_data(1, &"invalid", &nonce_program::ID);
+        let invalid_data = Account::new_data(1, &"invalid", &system_program::ID);
         assert_eq!(
             check_nonce_account(&invalid_data.unwrap(), &nonce_pubkey, &blockhash),
             Err(Box::new(CliError::InvalidNonce(
@@ -831,7 +831,7 @@ mod tests {
         let invalid_hash = Account::new_data(
             1,
             &NonceState::Initialized(NonceMeta::new(&nonce_pubkey), hash(b"invalid")),
-            &nonce_program::ID,
+            &system_program::ID,
         );
         assert_eq!(
             check_nonce_account(&invalid_hash.unwrap(), &nonce_pubkey, &blockhash),
@@ -841,7 +841,7 @@ mod tests {
         let invalid_authority = Account::new_data(
             1,
             &NonceState::Initialized(NonceMeta::new(&Pubkey::new_rand()), blockhash),
-            &nonce_program::ID,
+            &system_program::ID,
         );
         assert_eq!(
             check_nonce_account(&invalid_authority.unwrap(), &nonce_pubkey, &blockhash),
@@ -850,7 +850,7 @@ mod tests {
             ))),
         );
 
-        let invalid_state = Account::new_data(1, &NonceState::Uninitialized, &nonce_program::ID);
+        let invalid_state = Account::new_data(1, &NonceState::Uninitialized, &system_program::ID);
         assert_eq!(
             check_nonce_account(&invalid_state.unwrap(), &nonce_pubkey, &blockhash),
             Err(Box::new(CliError::InvalidNonce(
