+++ sdk/src/transaction.rs
@@ -3,10 +3,10 @@
 use crate::hash::Hash;
 use crate::instruction::{CompiledInstruction, Instruction, InstructionError};
 use crate::message::Message;
-use crate::nonce_instruction;
 use crate::pubkey::Pubkey;
 use crate::short_vec;
 use crate::signature::{KeypairUtil, Signature};
+use crate::system_instruction;
 use bincode::serialize;
 use std::result;
 
@@ -103,7 +103,8 @@ impl Transaction {
         nonce_authority_pubkey: &Pubkey,
         nonce_hash: Hash,
     ) -> Self {
-        let nonce_ix = nonce_instruction::nonce_advance(&nonce_account_pubkey, &nonce_authority_pubkey);
+        let nonce_ix =
+            system_instruction::nonce_advance(&nonce_account_pubkey, &nonce_authority_pubkey);
         instructions.insert(0, nonce_ix);
         Self::new_signed_with_payer(instructions, payer, signing_keypairs, nonce_hash)
     }
